Always flush within the write semaphore
It turns out that some streams are not safe to flush and write to concurrently. The STDIN stream of a child process happened to be the stream that we were writing to at a very high frequency when data corruption occurred. I saw cases both where buffers were transmitted twice, and in other cases buffers were just skipped.

Fixes #174