Close off race condition that led to unstable tests
The issue was that if the client had not yet exited the Writing semaphore by the time the server had processed the message and closed the connection, that the writing semaphore would throw ObjectDisposedException when we try to exit it. That would then be changed by MessageHandlerBase to an OperationCanceledException, and would propagate out from the `await this.TransmitAsync` call that's in our InvokeCoreAsync method out to our caller, entirely ignoring the `tcs.SetException(new ConnectionLostException())` we set elsewhere.

With this fix, we *guarantee* that the Task returned from InvokeCoreAsync will always change OperationCanceledException to ConnectionLostException if we're disconnected but the user hasn't canceled their token.