Trace message transmission failures
We didn't trace these before, and while `NotifyAsync` and other methods will propagate the exception to the caller, if the NotifyAsync call was made by a dynamically generated proxy in response to raising an event, the exception would be swallowed without a trace.

This also fixes invocation of the virtual `JsonRpc.OnRequestTransmissionAborted` method so that we only invoke it on actual transmission failures and not on all errors in receiving a response too.