Send error to client when server fails to serialize response
Previous to this change, a failure to serialize a response would lead the server to terminate the connection (since a response is required). With this change, we instead keep the connection going but send an error response to the client with the serialization error.

We also (still) log the serialization failure on the server.

Fixes #549